Update LORA_Crypt.php
Added a feature where the user can choose to encrypt/decrypt their files one at a time. 

Note: Threw in a sleep function to prevent an error after decrypting one file, it was doing unusual things when grabbing all the files in the directory again before it had refreshed, so now it just sleeps 8 seconds after doing so. Not the ideal solution, but for now it works.